requirements.txt
torch>=2.1.0
transformers>=4.40.0
tensorboard>=2.14.0
tqdm>=4.66.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
easyocr>=1.7.1